+++ src/main/java/grondag/adversity/niceblock/base/IFlowBlock.java
@@ -4,6 +4,8 @@ import net.minecraft.block.state.IBlockState;
 
 public interface IFlowBlock 
 {
-    public abstract float getRenderHeightFromState(IBlockState state);
+    /**
+     * Returns a value normalized from 0 to 15 irrespective of how many meta states are actually used  */
+    public abstract int getRenderHeightFromState(IBlockState state);
 
 }
